+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Fix python calls to patch_figure and revert_patch_figure to get api key from environment variable FIGPACK_API_KEY
+- Update manifest for patch-figure
 
 ## [0.3.11] - 2025-12-04